ran: format a SIB type as its number
Every log message and assertion that named a SIB had to cast it, each picking
its own integer type. Give sib_type a format_as, and drop the casts it makes
redundant.

+++ lib/du/du_high/du_manager/du_cell_manager.cpp
@@ -136,12 +136,9 @@ du_cell_manager::handle_cell_reconf_request(const du_cell_param_config_request& 
         // Increment value_tag with wrapping (5-bit field: 0-31).
         sib_it->value_tag = (sib_it->value_tag.value() + 1) % 32;
         si_updated        = true;
-        logger.info("Updated SIB{} in cell {} config, new value_tag={}",
-                    static_cast<int>(type),
-                    cell_index,
-                    sib_it->value_tag.value());
+        logger.info("Updated SIB{} in cell {} config, new value_tag={}", type, cell_index, sib_it->value_tag.value());
       } else {
-        logger.warning("Requested SIB{} update in cell {}, but entry not found.", static_cast<int>(type), cell_index);
+        logger.warning("Requested SIB{} update in cell {}, but entry not found.", type, cell_index);
       }
     }
   }
